adjusting build task for cypress
