+++ .github/workflows/main.yml
@@ -18,6 +18,7 @@ jobs:
       LOCALES: ${{ secrets.LOCALES }}
       NEXT_PUBLIC_SITE_URL: ${{ secrets.NEXT_PUBLIC_SITE_URL }}
       HASURA_API_URL: ${{ secrets.HASURA_API_URL }}
+      HASURA_ADMIN_SECRET: ${{ secrets.HASURA_ADMIN_SECRET }}
       ORG_SLUG: ${{ secrets.ORG_SLUG }}
       GOOGLE_CREDENTIALS_EMAIL: ${{ secrets.GOOGLE_CREDENTIALS_EMAIL }}
       GOOGLE_CREDENTIALS_PRIVATE_KEY: ${{ secrets.GOOGLE_CREDENTIALS_PRIVATE_KEY }}
@@ -51,7 +52,7 @@ jobs:
         with:
           # Disable running of tests within install job
           runTests: false
-          build: npm run build
+          build: npm run seed && npm run build:test
       - name: Save build folder
         uses: actions/upload-artifact@v2
         with:
