+++ script/shared.js
@@ -1,4 +1,5 @@
 const fetch = require('node-fetch');
+const faker = require('faker');
 
 const INSERT_LOCALE = `mutation FrontendInsertLocale($code: String!, $name: String!) {
   insert_locales_one(object: {code: $code, name: $name}, on_conflict: {constraint: locales_code_key, update_columns: code}) {
@@ -894,6 +895,291 @@ async function hasuraInsertTestArticle(params) {
   });
 }
 
+async function seedData(params) {
+  const deleteOrgResult = await hasuraRemoveOrganization({
+    url: params['url'],
+    adminSecret: params['adminSecret'],
+    slug: params['org']['slug'],
+  });
+  if (deleteOrgResult.errors) {
+    console.error(
+      params['adminSecret'],
+      'Error deleting organization: ',
+      deleteOrgResult.errors
+    );
+    // return orgResult.errors;
+  } else {
+    console.log('deleted organization:', deleteOrgResult);
+  }
+  const orgResult = await hasuraInsertOrganization({
+    url: params['url'],
+    adminSecret: params['adminSecret'],
+    name: params['org']['name'],
+    slug: params['org']['slug'],
+  });
+  if (orgResult.errors) {
+    console.error(
+      params['adminSecret'],
+      'Error creating organization: ',
+      orgResult.errors
+    );
+    return orgResult;
+  }
+  console.log('created organization:', orgResult);
+
+  let organizationID = orgResult.data.insert_organizations_one.id;
+  let name = params['org']['name'];
+
+  const localeResult = await hasuraInsertLocale({
+    url: params['url'],
+    adminSecret: params['adminSecret'],
+    name: 'English',
+    code: 'en-US',
+  });
+  if (localeResult.errors) {
+    console.error(
+      params['adminSecret'],
+      'Error creating locale: ',
+      localeResult.errors
+    );
+    return localeResult;
+  }
+  console.log('created en-US locale', localeResult);
+  let localeID = localeResult.data.insert_locales_one.id;
+
+  let orgLocaleObjects = [
+    {
+      locale_id: localeID,
+      organization_id: organizationID,
+    },
+  ];
+
+  console.log('orgLocaleObjects:', orgLocaleObjects);
+  let orgLocaleResult = await hasuraInsertOrgLocales({
+    url: params['url'],
+    adminSecret: params['adminSecret'],
+    orgLocales: orgLocaleObjects,
+  });
+  if (orgLocaleResult.errors) {
+    console.error(
+      params['adminSecret'],
+      'Error creating org locale: ',
+      orgLocaleResult.errors
+    );
+    return orgLocaleResult;
+  }
+  console.log('Setup locales for the organization:', orgLocaleResult);
+
+  let siteMetadata = {
+    color: 'colorone',
+    theme: 'styleone',
+    siteUrl: 'http://localhost:3000',
+    aboutCTA: 'Learn more',
+    aboutDek: `About the ${name} TK`,
+    aboutHed: 'Who We Are',
+    bodyFont: 'Domine',
+    shortName: name,
+    supportCTA: 'Donate',
+    supportDek: `${name} exists based on the support of our readers. Chip in today to help us continue delivering quality journalism.`,
+    supportHed: 'Support our work',
+    supportURL:
+      'https://tiny-news-collective.monkeypod.io/give/support-the-oaklyn-observer?secret=84fc2987ea6e8f11b8f4f8aca8b749d7',
+    footerTitle: 'Footer Title',
+    headingFont: 'Libre Franklin',
+    searchTitle: name,
+    primaryColor: '#de7a00',
+    twitterTitle: 'Twitter title',
+    facebookTitle: 'Facebook title',
+    homepageTitle: name,
+    membershipDek:
+      'Support great journalism by becoming a member for a low monthly price.',
+    membershipHed: 'Become a member',
+    newsletterDek: `Get the latest headlines from ${name} right in your inbox.`,
+    newsletterHed: 'Sign up for our newsletter',
+    donateBlockDek: 'Support our local journalism with a monthly pledge.',
+    donateBlockHed: 'Donate',
+    secondaryColor: '#002c57',
+    donationOptions:
+      '[{\n"amount": 5,\n"name": "Member",\n"description": "This is a description.",\n"cta": "Donate"\n},\n{\n"amount": 10,\n"name": "Supporter",\n"description": "This is a description.",\n"cta": "Donate"\n},\n{\n"amount": 20,\n"name": "Superuser",\n"description": "This is a description.",\n"cta": "Donate"\n}]',
+    footerBylineLink: 'http://localhost:3000',
+    footerBylineName: name,
+    searchDescription: 'Page description',
+    twitterDescription: 'Twitter description',
+    facebookDescription: 'Facebook description',
+    commenting: 'on',
+    advertisingHed: `Advertise with ${name}`,
+    advertisingDek:
+      'Want to reach our engaged, connected audience? Advertise within our weekly newsletter!',
+    advertisingCTA: 'Buy an advertisement',
+  };
+  let metadataResult = await hasuraUpsertMetadata({
+    url: params['url'],
+    adminSecret: params['adminSecret'],
+    organization_id: organizationID,
+    data: siteMetadata,
+    locale_code: 'en-US',
+    published: true,
+  });
+  if (metadataResult.errors) {
+    console.error(
+      params['adminSecret'],
+      'Error creating metadata: ',
+      metadataResult.errors
+    );
+    return metadataResult;
+  }
+  console.log('Setup metadata for the organization:', metadataResult);
+
+  let sectionResult = await hasuraInsertSections({
+    url: params['url'],
+    adminSecret: params['adminSecret'],
+    objects: [
+      {
+        organization_id: organizationID,
+        title: 'News',
+        slug: 'news',
+        published: true,
+        category_translations: {
+          data: {
+            locale_code: 'en-US',
+            title: 'News',
+          },
+          on_conflict: {
+            constraint: 'category_translations_locale_code_category_id_key',
+            update_columns: 'title',
+          },
+        },
+      },
+    ],
+  });
+  if (sectionResult.errors) {
+    console.error(
+      params['adminSecret'],
+      'Error creating section: ',
+      sectionResult.errors
+    );
+    return sectionResult;
+  }
+  console.log('Setup section for the organization:', sectionResult);
+  let categoryID = sectionResult.data.insert_categories.returning[0].id;
+  let largeLayoutResult = await hasuraUpsertHomepageLayout({
+    url: params['url'],
+    adminSecret: params['adminSecret'],
+    organization_id: organizationID,
+    name: 'Large Package Story Lead',
+    data:
+      '{ "subfeatured-top":"string", "subfeatured-bottom":"string", "featured":"string" }',
+  });
+  if (largeLayoutResult.errors) {
+    console.error(
+      params['adminSecret'],
+      'Error creating large package layout: ',
+      largeLayoutResult.errors
+    );
+    return largeLayoutResult;
+  }
+  console.log(
+    'Setup large package layout for the organization:',
+    largeLayoutResult
+  );
+
+  let bigLayoutResult = await hasuraUpsertHomepageLayout({
+    url: params['url'],
+    adminSecret: params['adminSecret'],
+    organization_id: organizationID,
+    name: 'Big Featured Story',
+    data: '{ "featured":"string" }',
+  });
+  if (bigLayoutResult.errors) {
+    console.error(
+      params['adminSecret'],
+      'Error creating big featured layout: ',
+      bigLayoutResult.errors
+    );
+    return bigLayoutResult;
+  }
+  console.log(
+    'Setup big featured layout for the organization:',
+    bigLayoutResult
+  );
+  let authorResult = await hasuraInsertOneAuthor({
+    url: params['url'],
+    orgSlug: params['orgSlug'],
+    first_names: faker.name.firstName(),
+    last_name: faker.name.lastName(),
+    title: faker.name.jobTitle(),
+    twitter: faker.internet.userName(),
+    bio: faker.name.jobDescriptor(),
+    slug: faker.lorem.slug(),
+    email: faker.internet.email(),
+  });
+  if (authorResult.errors) {
+    console.error(
+      params['orgSlug'],
+      'Error creating author: ',
+      authorResult.errors
+    );
+    return authorResult;
+  }
+  console.log('Setup author for the organization:', authorResult);
+  let authorID = authorResult.data.insert_authors_one.id;
+
+  let tagResult = await hasuraCreateOneTag({
+    url: params['url'],
+    orgSlug: params['orgSlug'],
+    slug: 'latest-news',
+    title: 'Latest News',
+    locale_code: 'en-US',
+  });
+  if (tagResult.errors) {
+    console.error(params['orgSlug'], 'Error creating tag:', tagResult.errors);
+    return tagResult;
+  }
+  console.log('Setup tag for the organization:', tagResult);
+  let tagID = tagResult.data.insert_tags_one.id;
+
+  let gdocResult = await hasuraInsertGoogleDoc({
+    url: params['url'],
+    orgSlug: params['orgSlug'],
+    document_id: '1LSyMzR1KxyKoml6q56DYQaxEV8Qm4EZo2y_xEFIkvGw',
+    locale_code: 'en-US',
+  });
+  if (gdocResult.errors) {
+    console.error(
+      params['adminSecret'],
+      'Error creating test google doc: ',
+      gdocResult.errors
+    );
+    return gdocResult;
+  }
+  console.log('Setup test google doc for the organization:', gdocResult);
+  let googleDocID = gdocResult.data.insert_google_documents_one.id;
+
+  let articleResult = await hasuraInsertTestArticle({
+    url: params['url'],
+    orgSlug: params['orgSlug'],
+    google_document_id: googleDocID,
+    locale_code: 'en-US',
+    category_id: categoryID,
+    slug: 'test-doc-for-article-features',
+    content: faker.lorem.paragraph(),
+    headline: faker.lorem.sentence(),
+    search_title: faker.lorem.sentence(),
+    search_description: faker.lorem.paragraph(),
+    author_id: authorID,
+    tag_id: tagID,
+  });
+  if (articleResult.errors) {
+    console.error(
+      params['orgSlug'],
+      'Error creating test article: ',
+      articleResult.errors
+    );
+    return articleResult;
+  }
+  console.log('Setup test article for the organization:', articleResult);
+  return orgResult;
+}
 async function fetchGraphQL(params) {
   let url;
   let orgSlug;
@@ -979,4 +1265,5 @@ module.exports = {
   hasuraCreateOneTag,
   fetchGraphQL,
   sanitizePath,
+  seedData,
 };
